analysis/llms/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Monitoring
starlette-prometheus==0.10.0

# AI/LLM dependencies
groq==0.11.0
google-generativeai==0.8.3

# HTTP client dependencies
httpx==0.25.2

# Utility dependencies
python-dotenv==1.0.0
